+++ Editor/EditorExample.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+using UnityEditor;
+
+namespace Hedron
+{
+    public class EditorExample : MonoBehaviour
+    {
+        private void Start() {
+            Debug.Log("Hello, world. It's me, EditorExample.Start()!");
+        }
+    }
+}
+