+++ Runtime/RuntimeExample.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace Hedron
+{
+    public class RuntimeExample : MonoBehaviour
+    {
+        private void Start() {
+            Debug.Log("Hello, world. It's me, RuntimeExample.Start()!");
+        }
+    }
+}