+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- A skeleton set of C# scripts for the Editor/ and Runtime/ environments.
 - A barebones package manifest to inform Unity of the package's information.
 - Initial `README` and `CHANGELOG` files to track project status.
 - Empty `LICENSE` file to remind me to eventually include an appropriate license.